\name{defNodesAtt}
\alias{defNodesAtt}
\title{Add nodes attributes}
\description{Internal}
\usage{
defNodesAtt(x, parent, ...)
}
\arguments{
  \item{x}{List of attributes}
  \item{parent}{xml parent}
  \item{\dots}{Further arguments to pass on to newXMLNode}
}
\author{George Vega Yon <george.vega@nodoschile.org>}