% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gexf_js_config}
\alias{gexf_js_config}
\alias{plot_gexfjs}
\title{Visualizing GEXF graph files using gexf-js (legacy)}
\usage{
gexf_js_config(
  dir,
  graphFile = "network.gexf",
  showEdges = TRUE,
  useLens = FALSE,
  zoomLevel = 0,
  curvedEdges = TRUE,
  edgeWidthFactor = 1,
  minEdgeWidth = 1,
  maxEdgeWidth = 50,
  textDisplayThreshold = 9,
  nodeSizeFactor = 1,
  replaceUrls = TRUE,
  showEdgeWeight = TRUE,
  showEdgeLabel = TRUE,
  sortNodeAttributes = TRUE,
  showId = TRUE,
  showEdgeArrow = TRUE,
  language = FALSE
)

plot_gexfjs(
  x,
  graphFile = "network.gexf",
  dir = tempdir(),
  overwrite = TRUE,
  httd.args = list(),
  copy.only = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory where files are copied (\code{tempdir()} by default).}

\item{graphFile}{Name of the GEXF file written to \code{dir}.}

\item{showEdges}{Logical scalar. Default state of the "show edges" button (nullable).}

\item{useLens}{Logical scalar. Default state of the "use lens" button (nullable).}

\item{zoomLevel}{Numeric scalar. Default zoom level. At zoom = 0, the graph
should fill a 800x700px zone}

\item{curvedEdges}{Logical scalar. False for curved edges, true for straight
edges this setting can't be changed from the User Interface.}

\item{edgeWidthFactor}{Numeric scalar. Change this parameter for wider or
narrower edges this setting can't be changed from the User Interface.}

\item{minEdgeWidth}{Numeric scalar.}

\item{maxEdgeWidth}{Numeric scalar.}

\item{textDisplayThreshold}{Numeric scalar.}

\item{nodeSizeFactor}{Numeric scalar. Change this parameter for smaller or
larger nodes this setting can't be changed from the User Interface.}

\item{replaceUrls}{Logical scalar. Enable the replacement of Urls by
Hyperlinks this setting can't be changed from the User Interface.}

\item{showEdgeWeight}{Logical scalar. Show the weight of edges in the list
this setting can't be changed from the User Interface.}

\item{showEdgeLabel}{Logical scalar.}

\item{sortNodeAttributes}{Logical scalar. Alphabetically sort node attributes.}

\item{showId}{Logical scalar. Show the id of the node in the list
this setting can't be changed from the User Interface.}

\item{showEdgeArrow}{Logical scalar. Show the edge arrows when the edge is
directed this setting can't be changed from the User Interface.}

\item{language}{Either \code{FALSE}, or a character scalar with any of the
supported languages.}

\item{x}{An object of class \code{gexf}.}

\item{overwrite}{Logical scalar. Overwrite existing files in \code{dir}.}

\item{httd.args}{Further arguments passed to \link[servr:httd]{servr::httd}.}

\item{copy.only}{Logical. When \code{TRUE}, skip launching the server.}

\item{...}{Further arguments passed to \link{gexf_js_config}.}
}
\description{
Uses the gexf-js JavaScript viewer by copying static files to a directory
and optionally launching a local web server via the \code{servr} package.  For
the modern htmlwidget-based renderer, use \link{plot.gexf} or \link{sigmajs}.
}
\details{
Currently, the only languages supported are:
German (de), English (en), French (fr), Spanish (es), Italian (it),
Finnish (fi), Turkish (tr), Greek (el), Dutch (nl)

The server is launched only when \code{interactive() == TRUE} and
\code{copy.only == FALSE}.
}
\examples{
if (interactive()) {
  path <- system.file("gexf-graphs/lesmiserables.gexf", package = "rgexf")
  graph <- read.gexf(path)
  plot_gexfjs(graph)
}

}
\references{
gexf-js project website \url{https://github.com/raphv/gexf-js}.
}
