\name{print.gexf}
\alias{print.gexf}
\title{Printing method for gexf class}
\description{To print graph gexf objects on screen or into a file}
\usage{
print.gexf(x, file, replace = F, ...)
}
\arguments{
  \item{x}{A gexf class object (builded from gexf())}
  \item{file}{Complete output path}
  \item{replace}{Logical. if TRUE replaces the target file (if exists)}
  \item{\dots}{Further arguments to pass on print}
}
\author{George Vega Yon <george.vega@nodoschile.org>}

