\name{processEdgeList}
\alias{processEdgeList}
\title{Decompose an edge list}
\description{Generates two data frames (nodes and edges) from a edge list}
\usage{
processEdgeList(x)
}
\arguments{
  \item{x}{A matrix or data frame structured as an edge list}
}
\details{
  \code{processEdgeList} transforms it's input into a two element list containing a data frame of nodes (with columns 'id' and 'label') and another one, numeric, of edges (with columns 'source' and 'target'), with the last one based on autogenerated nodes ids. 
}
\value{
  A list containing two data frames.
}
\author{George Vega Yon \email{george.vega@nodoschile.org}}
\examples{
  edgelist <- matrix(
    c("matthew","john","max","stephen","matthew","stephen"),
    byrow=TRUE, ncol=2)
  
  processEdgeList(edgelist)
}