% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs.r
\name{checkTimes}
\alias{checkTimes}
\title{Checks for correct time format}
\usage{
checkTimes(x, format = "date")
}
\arguments{
\item{x}{A string or vector char}

\item{format}{String, can be \dQuote{date}, \dQuote{dateTime},
\dQuote{float}}
}
\value{
Logical.
}
\description{
Checks time
}
\examples{

  test <- c("2012-01-17T03:46:41", "2012-01-17T03:46:410")
  checkTimes(test, format="dateTime")
  checkTimes("2012-02-01T00:00:00", "dateTime")
}
\author{
George Vega Yon 

Jorge Fabrega Lacoa
}
\keyword{utilities}

