\name{checkTimes}
\alias{checkTimes}
\title{Checks for correct time format}
\description{Checks time}
\usage{
checkTimes(x, format = "date")
}
\arguments{
  \item{x}{A string char}
  \item{format}{format, could be "date", "dateTime", "float"}
}
\author{George Vega Yon \email{george.vega@nodoschile.org}}