% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgexf-package.R
\docType{data}
\name{followers}
\alias{followers}
\title{Edge list with attributes}
\format{
A data frame containing 6065 observations.
}
\source{
Fabrega and Paredes (2012): \dQuote{La politica en 140 caracteres}
en Intermedios: medios de comunicacion y democracia en Chile. Ediciones UDP
}
\description{
Sample of accounts by December 2011.
}
\keyword{datasets}
