% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.r
\name{igraph.to.gexf}
\alias{gexf.to.igraph}
\alias{igraph.to.gexf}
\title{Converting between \code{gexf} and \code{igraph} classes}
\usage{
igraph.to.gexf(igraph.obj, position = NULL)

gexf.to.igraph(gexf.obj)
}
\arguments{
\item{igraph.obj}{An object of class \code{igraph}.}

\item{position}{A three-column data-frame with XYZ coords.}

\item{gexf.obj}{An object of class \code{gexf}.}
}
\value{
\itemize{ \item For \code{igraph.to.gexf} : \code{gexf} class object
\item For \code{gexf.to.igraph} : \code{igraph} class object }
}
\description{
Converts objects between \code{gexf} and \code{igraph} objects keeping
attributes, edge weights and colors.
}
\details{
If the position argument is not \code{NULL}, the new \code{gexf} object will
include the \code{position} viz-attribute.
}
\examples{

 \dontrun{
 
  # Running demo
  demo(gexfigraph)
 
  fn <- system.file("gexf-graphs/lesmiserables.gexf", package = "rgexf")
  gexf1 <- read.gexf(fn)
  igraph1 <- gexf.to.igraph(gexf1)
  gexf2 <- igraph.to.gexf(igraph1)
 }
}
\author{
George Vega Yon \email{g.vegayon@gmail.com}
}
\seealso{
\code{\link{layout}}
}
\keyword{manip}

