% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gexf.r
\name{new.gexf.graph}
\alias{new.gexf.graph}
\title{Build an empty \code{gexf} graph}
\usage{
new.gexf.graph(
  defaultedgetype = "undirected",
  meta = list(creator = "NodosChile", description =
    "A graph file writing in R using 'rgexf'", keywords =
    "gexf graph, NodosChile, R, rgexf")
)
}
\arguments{
\item{defaultedgetype}{\dQuote{directed}, \dQuote{undirected},
\dQuote{mutual}}

\item{meta}{A List. Meta data describing the graph}
}
\value{
A \code{gexf} object.
}
\description{
Builds an empty \code{gexf} object containing all the class's attributes.
}
\examples{

  \dontrun{
  demo(gexfbuildfromscratch)
  }

}
\references{
The GEXF project website: http://gexf.net
}
\author{
George Vega Yon

Jorge Fabrega Lacoa
}
\keyword{manip}
