% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexf-js.r
\name{gexf_js_make_config}
\alias{gexf_js_make_config}
\alias{plot.gexf}
\title{Visualizing GEXF graph files using gexf-js}
\usage{
gexf_js_make_config(dir, graphFile = "network.gexf", config = list(showEdges
  = "true", useLens = "false", zoomLevel = "0", curvedEdges = "true",
  edgeWidthFactor = "1", minEdgeWidth = "1", maxEdgeWidth = "50",
  textDisplayThreshold = "9", nodeSizeFactor = "1", replaceUrls = "true",
  showEdgeWeight = "true", showEdgeLabel = "true", sortNodeAttributes = "true",
  showId = "true", showEdgeArrow = "true", language = "false"))

\method{plot}{gexf}(x, y = NULL, graphFile = "network.gexf",
  dir = tempdir(), overwrite = TRUE, httd.args = list(),
  copy.only = FALSE, ...)
}
\arguments{
\item{dir}{Directory where the files will be copied (tempdir() by default).}

\item{graphFile}{Name of the gexf file.}

\item{config}{List of parameters to be written in \code{config.js}.}

\item{x}{An object of class \code{gexf}.}

\item{y}{Ignored.}

\item{overwrite}{Logical scalar. When \code{TRUE}, the default, the function
will overwrite all files copied from the template on the destination directory
as specified by \code{dir}.}

\item{httd.args}{Further arguments to be passed to \code{\link[servr:httd]{httd}}
from the \CRANpkg{servr} package.}

\item{copy.only}{Logical scalar. When FALSE, the default, the function
will make a call to \code{servr::httd}.}

\item{...}{Further arguments passed to \code{gexf_js_make_config}}
}
\description{
Using the gexf-js, a JavaScript GEXF viewer, this function allows you to
visualize your GEXF on the browser. The function essentially copies a template
website, the GEXF file, and sets up a configuration file. By default, the
function then starts a webserver using the \code{servr} R package.
}
\details{
The files are copied directly from
\Sexpr[results=text]{system.file("gexf-js", package="rgexf")}. And the 
parameters are set up by modifying the following template file:
\Sexpr[results=text]{system.file("gexf-js/config.js.template", package="rgexf")}
}
\examples{
\dontrun{

path <- system.file("gexf-graphs/lesmiserables.gexf", package="rgexf")
graph <- read.gexf(path)
plot(graph)

}

}
\references{
gexf-js project website \url{https://github.com/raphv/gexf-js}.
}
