% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gexf.R
\name{read.gexf}
\alias{read.gexf}
\title{Reads gexf (.gexf) file}
\usage{
read.gexf(x)
}
\arguments{
\item{x}{String. Path to the gexf file.}
}
\value{
A \code{gexf} object.
}
\description{
\code{read.gexf} reads gexf graph files and imports its elements as a
\code{gexf} class object
}
\note{
By the time attributes and viz-attributes aren't supported.
}
\examples{

  fn <- system.file("gexf-graphs/lesmiserables.gexf", package = "rgexf")
  mygraph <- read.gexf(fn)

}
\references{
The GEXF project website: http://gexf.net
}
\author{
George Vega Yon

Jorge Fabrega Lacoa
}
\keyword{IO}
