\name{rgexf-package}
\alias{rgexf-package}
\alias{rgexf}
\docType{package}
\title{An R library tu build GEXF graph files}
\description{Builds GEXF files}
\details{
\tabular{ll}{
Package: \tab rgexf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-28\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
George Vega Yon <george.vega@nodoschile.org>

Maintainer: George Vega Yon <george.vega@nodoschile.org>
}
