\name{rgexf-package}
\alias{rgexf-package}
\alias{rgexf}
\alias{gephi}
\docType{package}
\title{
  An R library to build GEXF graph files
}
\description{
  rgexf packages provides of a collection of functions to generate GEXF and read graph files supported on GEPHI.
  
  It supports building dynamic graphs and adding attributes to edges and nodes, including visual attributes such as color, shape and size.
  
  Please visit the project home for more information:
    \url{https://bitbucket.org/gvegayon/rgexf}.
} 
\details{
  \tabular{ll}{
    Package: \tab rgexf\cr
    Type: \tab Package\cr
    Version: \tab 0.12.12\cr
    Date: \tab 2012-12-26\cr
    License: \tab GPL version 2 or later\cr
  }
}
\author{George Vega Yon \email{george.vega@nodoschile.org},

Jorge Fabrega Lacoa \email{jorge.fabrega@nodoschile.org}}
\note{
  See the GEXF primer for details on the GEXF graph format:
  \url{http://gexf.net/1.2draft/gexf-12draft-primer.pdf}
}
\references{
  \itemize{
    \item rgexf project site: \url{https://bitbucket.org/gvegayon/rgexf}
    \item Gephi project site: \url{https://gephi.org/}
    \item GEXF project site: \url{http://gexf.net/}
    \item Nodos Chile project site (real life examples): \url{http://nodoschile.org/}
  }
}
\examples{
  \dontrun{
    demo(gexf)
    demo(gexffull)
    demo(gexftwitter)
  }
}
\keyword{ package }
