\name{twitteraccounts}
\docType{data}
\alias{twitteraccounts}
\title{Twitter accounts of Chilean Politicians and Journalists (sample)}
\description{
  Sample of accounts by december 2011.
}
\usage{data(twitteraccounts)}
\format{A data frame containing 148 observations.}
\source{
  Fabrega and Paredes (2012): ``La politica en 140 caracteres'' en Intermedios: medios de comunicacion y democracia en Chile. Ediciones UDP}
\keyword{ datasets }